requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
psycopg2==2.9.6
djangorestframework==3.14.0
Faker==19.12.0
django-redis==5.4.0